# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Caffe-Backend",
        "version": "0.1"
    },
    "paths": {
        "/api/entrada-produto": {
            "get": {
                "tags": [
                    "Entrada de Produto"
                ],
                "summary": "Listar entradas de produtos",
                "operationId": "0fcfca1b8f09c2c0f8f2e4653e1e0c82",
                "parameters": [
                    {
                        "name": "idProduto",
                        "in": "query",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de entradas de produtos"
                    },
                    "400": {
                        "description": "Erro ao buscar entrada de produtos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Entrada de Produto"
                ],
                "summary": "Registrar entrada de produto",
                "operationId": "3da8d2c60f20b4bdbea5d7220060cb54",
                "requestBody": {
                    "description": "Informações da entrada de produto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idProduto",
                                    "Quantidade",
                                    "Preco",
                                    "Data",
                                    "Validade",
                                    "Informacoes_adicionais"
                                ],
                                "properties": {
                                    "idProduto": {
                                        "type": "integer"
                                    },
                                    "Quantidade": {
                                        "type": "integer"
                                    },
                                    "Preco": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "Data": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "Validade": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "Informacoes_adicionais": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entrada de produto cadastrada com sucesso"
                    },
                    "400": {
                        "description": "Erro ao cadastrar entrada de produto"
                    }
                }
            }
        },
        "/api/entrada-produto/{id}": {
            "get": {
                "tags": [
                    "Entrada de Produto"
                ],
                "summary": "Mostrar detalhes da entrada de produto",
                "operationId": "cda5a4778fb59aabb77a49d60dd01904",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da entrada de produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os detalhes da entrada de produto"
                    },
                    "400": {
                        "description": "Erro ao buscar entrada de produto"
                    }
                }
            },
            "put": {
                "tags": [
                    "Entrada de Produto"
                ],
                "summary": "Atualizar entrada de produto",
                "operationId": "ad8bb0b99a87cbbfc0a4beeac1137308",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da entrada de produto a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Informações da entrada de produto a serem atualizadas",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idProduto": {
                                        "type": "integer"
                                    },
                                    "Quantidade": {
                                        "type": "integer"
                                    },
                                    "Preco": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "Data": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "Validade": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "Informacoes_adicionais": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entrada de produto atualizada com sucesso"
                    },
                    "400": {
                        "description": "Erro ao atualizar entrada de produto"
                    }
                }
            }
        },
        "/api/itens-mesa": {
            "get": {
                "tags": [
                    "Itens da Mesa"
                ],
                "summary": "Listar itens da mesa",
                "operationId": "d95cc60d7c93f9a942f33ade5384c78d",
                "parameters": [
                    {
                        "name": "idMesa",
                        "in": "query",
                        "description": "ID da mesa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de itens da mesa"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Itens da Mesa"
                ],
                "summary": "Adicionar item à mesa",
                "operationId": "3fad498df8be3051fad392e110eb1d10",
                "requestBody": {
                    "description": "Informações do novo item da mesa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idMesa",
                                    "Quantidade",
                                    "Total_soma",
                                    "idProduto"
                                ],
                                "properties": {
                                    "idMesa": {
                                        "type": "integer"
                                    },
                                    "Quantidade": {
                                        "type": "integer"
                                    },
                                    "Total_soma": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "idProduto": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item da mesa adicionado com sucesso"
                    },
                    "422": {
                        "description": "Erros de validação"
                    }
                }
            }
        },
        "/api/itens-mesa/{id}": {
            "put": {
                "tags": [
                    "Itens da Mesa"
                ],
                "summary": "Atualizar item da mesa",
                "operationId": "f89f5929c1bd1415b5dcb2f9f8cab448",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do item da mesa a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Informações do item da mesa a serem atualizadas",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idMesa": {
                                        "type": "integer"
                                    },
                                    "Quantidade": {
                                        "type": "integer"
                                    },
                                    "Total_soma": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "idProduto": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item da mesa atualizado com sucesso"
                    },
                    "404": {
                        "description": "Item da mesa não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
...
    }
}